# Conflicts:
#	Ghostbait/Ghostbait/Ghostbait.vcxproj
#	Ghostbait/Ghostbait/Ghostbait.vcxproj.filters
#	Ghostbait/Ghostbait/Header Files/HexGrid.h
#	Ghostbait/Ghostbait/Header Files/Transform.h
#	Ghostbait/Ghostbait/Source Files/Console.cpp
#	Ghostbait/Ghostbait/Source Files/HexGrid.cpp
#	Ghostbait/Ghostbait/Source Files/Main.cpp
#	Ghostbait/Ghostbait/Source Files/Player.cpp
#	Ghostbait/Ghostbait/Source Files/SceneManager.cpp

+++ Ghostbait/Ghostbait/Source Files/SpatialPartition.cpp
@@ -156,7 +156,7 @@ std::vector<uint32_t> SpatialPartition::Hash(const AABB aabb) {
 	}
 	
 	int index;
-	for(int point = 0; point < points.size(); ++point) {
+	for(size_t point = 0; point < points.size(); ++point) {
 		index = Hash(points[point]);
 		bool found = false;
 		//for (unsigned int exist = 0; exist < indicies.size(); ++exist) {
